add placeholder profile settings page, update auth token expiration to 30 days
