+++ smart-hydration-app/app/(main)/settings/profile.tsx
@@ -10,17 +10,18 @@ import { ISettingsSection } from "@/interfaces/settings";
 
 const settingsList: ISettingsSection[] = [
     {
-        title: "Account",
+        title: "",
         data: [
             {
-                name: "Profile",
+                name: "Change email",
                 component: (name, isFirst, isLast) => {
                     const router = useRouter();
                     return (
                         <OptionBlock
                             isLast={isLast}
+                            isFirst={isFirst}
                             text={name}
-                            onPress={() => router.navigate("settings/profile")}
+                            // onPress={() => router.navigate("add-device-modal")}
                             icon={
                                 <Feather name="user" size={18} color="gray" />
                             }
@@ -29,45 +30,17 @@ const settingsList: ISettingsSection[] = [
                 },
             },
             {
-                name: "User Mode",
+                name: "Change password",
                 component: (name, isFirst, isLast) => {
                     const router = useRouter();
                     return (
                         <OptionBlock
                             isLast={isLast}
+                            isFirst={isFirst}
                             text={name}
-                            onPress={() => router.navigate("settings/mode")}
+                            // onPress={() => router.navigate("add-device-modal")}
                             icon={
-                                <MaterialCommunityIcons
-                                    name="cards-playing-heart-multiple-outline"
-                                    size={16}
-                                    color="gray"
-                                />
-                            }
-                        />
-                    );
-                },
-            },
-        ],
-    },
-    {
-        title: "Appearance",
-        data: [
-            {
-                name: "Theme",
-                component: (name, isFirst, isLast) => {
-                    const router = useRouter();
-                    return (
-                        <OptionBlock
-                            isLast={isLast}
-                            text={name}
-                            onPress={() => router.navigate("settings/theme")}
-                            icon={
-                                <Ionicons
-                                    name="color-palette"
-                                    size={19}
-                                    color="gray"
-                                />
+                                <Feather name="user" size={18} color="gray" />
                             }
                         />
                     );
@@ -84,7 +57,7 @@ export default function SettingsModal() {
 
     return (
         <View
-            className="flex flex-1 justify-between mx-4"
+            className="flex flex-1 justify-between mx-4 mt-4"
             style={{
                 paddingBottom: insets.bottom + 20,
             }}
@@ -98,13 +71,6 @@ export default function SettingsModal() {
                         index == section.data.length - 1,
                     )
                 }
-                renderSectionHeader={({ section }) => (
-                    <View className="bg-gray-100 dark:bg-neutral-900 py-4 px-4 rounded-t-xl mt-6">
-                        <Text className="font-bold dark:text-white">
-                            {section.title}
-                        </Text>
-                    </View>
-                )}
                 keyExtractor={(item) => `settings-${item.name}`}
                 stickySectionHeadersEnabled={false}
             />
